requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
loguru>=0.7.2
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
google-genai>=0.3.0
oss2>=2.18.0
